🐛 fix(git): Build warp-input on push to main (#57)
## Description

**What issue are you solving (or what feature are you adding) and how
are you doing it?**

There is a bug whereby `netris/warp-input` does not build and deploy a
new nightly docker image when a pr is merged. This happens when there
are changes in `bin/input/**` directory, and this is not the expected
behaviour. So i added the directory back into the `warp-input.yml`
workflow file.